(fix|COS-3846): Allow user to change period for newly created service line items or items for draft contracts
